HPS-27 refactor: enhance error handling structures
