+++ src/main/java/com/example/colaba/exception/GlobalExceptionHandler.java
@@ -1,13 +1,15 @@
 package com.example.colaba.exception;
 
+import com.example.colaba.exception.common.DuplicateEntityException;
+import com.example.colaba.exception.common.NotFoundException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
-    @ExceptionHandler(UserNotFoundException.class)
-    public ResponseEntity<String> handleUserNotFound(UserNotFoundException e) {
+    @ExceptionHandler(NotFoundException.class)
+    public ResponseEntity<String> handleNotFound(NotFoundException e) {
         return ResponseEntity.notFound().build();
     }
