--- src/main/java/com/example/colaba/exception/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.colaba.exception;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
