+++ src/main/java/com/example/colaba/exception/common/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.example.colaba.exception.common;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
