+++ src/main/java/com/example/colaba/exception/task/TaskNotFoundException.java
@@ -1,7 +1,9 @@
 package com.example.colaba.exception.task;
 
-public class TaskNotFoundException extends RuntimeException {
+import com.example.colaba.exception.common.NotFoundException;
+
+public class TaskNotFoundException extends NotFoundException {
     public TaskNotFoundException(String message) {
-        super(message);
+        super("Task not found: " + message);
     }
 }
