+++ src/main/java/com/example/colaba/exception/user/UserNotFoundException.java
@@ -1,0 +1,9 @@
+package com.example.colaba.exception.user;
+
+import com.example.colaba.exception.common.NotFoundException;
+
+public class UserNotFoundException extends NotFoundException {
+    public UserNotFoundException(String message) {
+        super("User not found: " + message);
+    }
+}
